+++ CHANGELOG.md
@@ -25,7 +25,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds an `onSubmit` prop to the `Form` component.
 - Adds the `renderHeaderElement` prop to the `TemplateAppContainer` component. This prop is used to control whether the `TemplateAppContainer` component should render its own `<header>` HTML element through its `header` prop, or let the user pass in their own component that renders the `<header>` HTML element.
 - Updates the `Fieldset` and `RadioGroup` so the `children` prop is declared through `React.PropsWithChildren` rather than in their respective prop interfaces.
-- Adds snapshot tests for the `Accordian` component, as well as `id` props to the components within the `Accordian` so snapshot tests pass.
+- Adds snapshot tests for the `Accordion` component, as well as `id` props to the components within the `Accordion` so snapshot tests pass.
 
 ### Fixes